chore: refactor error handling
