+++ crates/yttrium/Cargo.toml
@@ -20,6 +20,7 @@ wasm = [
     "dep:wasm-bindgen-futures",
     "alloy/wasm-bindgen",
     "dep:tsify-next",
+    "dep:derive_jserror",
     # "getrandom/wasm_js", # for getrandom v0.3.*
     "getrandom/js",
     "uuid/js",
@@ -97,6 +98,9 @@ tracing = { version = "0.1.40", default-features = false, features = [
 
 # Networking
 reqwest.workspace = true
+
+# Tiny dependency, could copy source. Pinning to exact version to avoid supply-chain attacks.
+derive_jserror = { version = "=0.1.0", optional = true, default-features = false }
 
 [dev-dependencies]
 # mocking
