+++ crates/yttrium/src/chain_abstraction/mod.rs
@@ -6,6 +6,7 @@ pub mod error;
 pub mod l1_data_fee;
 pub mod local_fee_acc;
 pub mod pulse;
+pub mod send_transaction;
 pub mod spawn;
 pub mod ui_fields;
