+++ crates/yttrium/src/wasm_compat.rs
@@ -17,7 +17,6 @@ use {
             ui_fields::UiFields,
         },
     },
-    alloy::primitives::{Address, PrimitiveSignature},
     std::time::Duration,
     wasm_bindgen::prelude::*,
 };
@@ -42,14 +41,9 @@ impl Client {
         chain_id: String,
         from: String,
         call: Call,
-    ) -> Result<PrepareResponse, JsValue> {
+    ) -> Result<PrepareResponse, JsError> {
         self.inner
-            .prepare(
-                chain_id,
-                from.parse::<Address>()
-                    .map_err(|e| JsValue::from_str(&e.to_string()))?,
-                call,
-            )
+            .prepare(chain_id, from.parse()?, call)
             .await
             .map_err(Into::into)
     }
@@ -70,15 +64,9 @@ impl Client {
         from: String,
         call: Call,
         local_currency: Currency,
-    ) -> Result<PrepareDetailedResponse, JsValue> {
+    ) -> Result<PrepareDetailedResponse, JsError> {
         self.inner
-            .prepare_detailed(
-                chain_id,
-                from.parse::<Address>()
-                    .map_err(|e| JsValue::from_str(&e.to_string()))?,
-                call,
-                local_currency,
-            )
+            .prepare_detailed(chain_id, from.parse()?, call, local_currency)
             .await
             .map_err(Into::into)
     }
@@ -126,17 +114,19 @@ impl Client {
         ui_fields: UiFields,
         route_txn_sigs: Vec<String>,
         initial_txn_sig: String,
-    ) -> Result<ExecuteDetails, JsValue> {
+    ) -> Result<ExecuteDetails, JsError> {
         self.inner
             .execute(
                 ui_fields,
-                route_txn_sigs
-                    .iter()
-                    .map(|s| s.parse::<PrimitiveSignature>().unwrap())
-                    .collect(),
-                initial_txn_sig
-                    .parse::<PrimitiveSignature>()
-                    .map_err(|e| e.to_string())?,
+                {
+                    // TODO refactor to use try_collect() when it's stable
+                    let mut sigs = Vec::with_capacity(route_txn_sigs.len());
+                    for result in route_txn_sigs.iter().map(|s| s.parse()) {
+                        sigs.push(result?);
+                    }
+                    sigs
+                },
+                initial_txn_sig.parse()?,
             )
             .await
             .map_err(Into::into)
@@ -147,16 +137,11 @@ impl Client {
         chain_id: &str,
         token: String,
         owner: String,
-    ) -> Result<String, String> {
-        Ok(self
-            .inner
-            .erc20_token_balance(
-                chain_id,
-                token.parse::<Address>().map_err(|e| e.to_string())?,
-                owner.parse::<Address>().map_err(|e| e.to_string())?,
-            )
+    ) -> Result<String, JsError> {
+        self.inner
+            .erc20_token_balance(chain_id, token.parse()?, owner.parse()?)
             .await
-            .map_err(|e| e.to_string())?
-            .to_string())
+            .map_err(Into::into)
+            .map(|balance| balance.to_string())
     }
 }
